# Optimize allLocalesBarrier by reducing communication