--- a/modules/standard/Barriers.chpl
+++ b/modules/standard/Barriers.chpl
@@ -259,7 +259,7 @@ module Barriers {
         if myc<=1 {
           if hackIntoCommBarrier {
             extern proc chpl_comm_barrier(msg: c_string);
-            chpl_comm_barrier("local barrier call".localize().c_str());
+            chpl_comm_barrier(c"local barrier call");
           }
           const alreadySet = done.testAndSet();
           if boundsChecking && alreadySet {