--- a/test/parallel/taskPar/sungeun/barrier/commDiags.na-none.good
+++ b/test/parallel/taskPar/sungeun/barrier/commDiags.na-none.good
@@ -22,3 +22,9 @@ sync remote test split phase
 (get = 2, execute_on = 2, execute_on_fast = 1)
 (get = 2, execute_on = 2, execute_on_fast = 1)
 (get = 2, execute_on = 2, execute_on_fast = 1)
+allLocalesBarrier test basic
+(execute_on_nb = 4)
+(execute_on_fast = 1)
+(execute_on_fast = 1)
+(execute_on_fast = 1)
+(execute_on_fast = 1)