@ronawho 2021-07-07
The barrier uses atomics, which fence the cache so the cache does not help at all for this.